Rename mapper to converter
