+++ sonar-apple-plugin/src/main/java/fr/insideapp/sonarqube/apple/xcode/warnings/XcodeWarningExtensionProvider.java
@@ -18,7 +18,7 @@
 package fr.insideapp.sonarqube.apple.xcode.warnings;
 
 import fr.insideapp.sonarqube.apple.commons.ExtensionProvider;
-import fr.insideapp.sonarqube.apple.xcode.warnings.mapper.XcodeWarningsMapper;
+import fr.insideapp.sonarqube.apple.xcode.warnings.converter.XcodeWarningsConverter;
 import fr.insideapp.sonarqube.apple.xcode.warnings.parser.XcodeWarningParser;
 import fr.insideapp.sonarqube.objc.issues.warnings.XcodeWarningObjectiveCRulesDefinition;
 import fr.insideapp.sonarqube.swift.issues.warnings.XcodeWarningSwiftRulesDefinition;
@@ -35,7 +35,7 @@ public final class XcodeWarningExtensionProvider implements ExtensionProvider {
             XcodeWarningObjectiveCRulesDefinition.class,
             XcodeWarningSwiftRulesDefinition.class,
             XcodeWarningParser.class,
-            XcodeWarningsMapper.class,
+            XcodeWarningsConverter.class,
             XcodeWarningsSensor.class
         );
     }
