+++ sonar-apple-plugin/src/test/resources/xcode/warnings/converter/warningSummary_noLocation.json
